requirements.txt
discord.py
python-decouple
pyprofiles==0.0.2